+++ src/store/index.js
@@ -8,7 +8,9 @@ export default new Vuex.Store({
     drawerState: true,
     basemapState: "Topo",
     streamgageState: false,
+    radarState: false,
     nfhlState: false,
+    fwwState: false,
     currentZoomState: 4,
   },
   mutations: {
@@ -21,8 +23,14 @@ export default new Vuex.Store({
     getStreamgageState(state, streamgagePicked) {
       state.streamgageState = streamgagePicked;
     },
+    getRadarState(state, radarPicked) {
+      state.radarState = radarPicked;
+    },
     getNfhlState(state, nfhlPicked) {
       state.nfhlState = nfhlPicked;
+    },
+    getFwwState(state, fwwPicked) {
+      state.fwwState = fwwPicked;
     },
     getCurrentZoomState(state, currentZoom) {
       state.currentZoomState = currentZoom;
@@ -32,7 +40,9 @@ export default new Vuex.Store({
     drawerState: (state) => state.drawerState,
     basemapState: (state) => state.basemapState,
     streamgageState: (state) => state.streamgageState,
+    radarState: (state) => state.radarState,
     nfhlState: (state) => state.nfhlState,
+    fwwState: (state) => state.fwwState,
     currentZoomState: (state) => state.currentZoomState,
   },
   actions: {},