+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Uncategorized icon to map and legend
 - Added NFHL layer and legend for NFHL sublayer
 - Added No active flooding dialog
+- Added NWS Radar layer
 
 ### Changed
 
 - RP icons visible in legend only if on the map
+- NWIS streamgages shown above RPs when layer is on
 
 ### Fixed